make snapshots replication aware

+++ snapshots/process-snaps.sh
@@ -55,11 +55,35 @@ for pool in $pools; do
         
         for job in $jobs; do
             zfsfolder=`echo $job|${SED} 's,%,/,g'`
+            # Make sure we are not a replication target
+            case $(replication_source $zfsfolder) in 
+                'ERROR')
+                    # zfsfolder does not exist
+                    snap_this_folder='false'
+                    ;;
+                'NONE')
+                    # No replication
+                    snap_this_folder='true'
+                    ;;
+                "$pool")
+                    # This is a source folder
+                    snap_this_folder='true'
+                    ;;
+                *)
+                    # This is a target folder
+                    snap_this_folder='false'
+                    ;;
+            esac
+            if [ "$snap_this_folder" == 'false' ]; then
+                # We should not snapshot this folder
+                continue
+            fi
+
             keepcount=`cat $jobfolder/$snaptype/$job`
             now=`${DATE} +%F_%H:%M%z`
             stamp="${snaptype}_${now}"
             if [ "${keepcount:0:1}" != "x" ]; then
-                zfs snapshot ${zfsfolder}@${stamp} 2> ${TMP}/process_snap_$$ ; result=$?
+                                zfs snapshot ${zfsfolder}@${stamp} 2> ${TMP}/process_snap_$$ ; result=$?
                 if [ "$result" -ne "0" ]; then
                     error "Failed to create snapshot ${zfsfolder}@${stamp}" ${TMP}/process_snap_$$
                     rm ${TMP}/process_snap_$$
